수정된 ERD 반영한 Domain Commit
수정된 ERD(20250720) Version
Domain(DTO) 수정.
+ Vehicle Domain 삭제되었지만 Post Method 예시 위해 관련 service 등 미삭제.

+++ src/main/java/com/ionecar/domain/Vehicle.java
@@ -1,18 +1,14 @@
 package com.ionecar.domain;
 
 import lombok.Data;
-import java.math.BigDecimal;
 
 @Data
 public class Vehicle {
-
     private String vehicleId;
     
     private String modelName;
     private String manufacturer;
-    private BigDecimal basePrice;
+    private long basePrice;
     private int modelYear;
     private String fuelType;
-
 }
-
